requirements.txt
requests==2.31.0
web3==6.15.1
eth-account==0.10.0
python-dotenv==1.0.0
pydantic==2.6.1
